Refactor(renderer): Rename reactive-data.js to attribute-binding.js
